+++ .storybook/preview.js
@@ -1,3 +1,5 @@
+// import '!style-loader!css-loader!sass-loader!../src/assets/scss/main.scss';
+import '../src/assets/css/main.css';
 
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },