+++ src/install.ts
@@ -4,7 +4,7 @@ import components, { PhotonComponentKeys } from './stories'
 import Vue, { Component } from 'vue';
 import { DefaultData, DefaultMethods, DefaultComputed, DefaultProps } from 'vue/types/options';
 import VueCompositionApi from '@vue/composition-api';
-import './assets/css/main.css'
+import './assets/scss/main.scss'
 
 Vue.use(VueCompositionApi);
 
@@ -13,7 +13,12 @@ interface options {
 }
 
 export default {
-  install(Vue: { component: (arg0: string, arg1: Component<DefaultData<never>, DefaultMethods<never>, DefaultComputed, DefaultProps>) => void; }, options?: options): void {
+  install(Vue: { component: (
+    arg0: string, 
+    arg1: Component<DefaultData<never>, DefaultMethods<never>, DefaultComputed, DefaultProps>
+  ) => void; }, 
+  options?: options
+  ): void {
     for (const key in components) {
       const _key: PhotonComponentKeys = key
       Vue.component(key, components[_key]);